% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedRlabD-package.R
\docType{package}
\name{AdvancedRlabD}
\alias{AdvancedRlabD-package}
\alias{AdvancedRlabD}
\title{AdvancedRlabD}
\description{
This package contains functions and datasets developed for the
Advanced R Lab 3D course.
}
\details{
The package provides the \code{dijkstra()} function for calculating
shortest distances in a weighted graph using Dijkstra's algorithm,
and the \code{euclidean()} function for calculating the greatest
common divisor using the Euclidean algorithm.

The package also includes the \code{wiki_graph} dataset, which
contains the edges and weights of a graph used with the
\code{dijkstra()} function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sheethal098/AdvancedRlabD}
  \item Report bugs at \url{https://github.com/sheethal098/AdvancedRlabD/issues}
}

}
\author{
\strong{Maintainer}: Sheethal Lakshmana \email{shehe874@student.liu.se}

Authors:
\itemize{
  \item Sheethal Lakshmana \email{shehe874@student.liu.se}
  \item Mariam Yayloyan \email{marya117@student.liu.se}
}

}
