% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wikipedia graph dataset}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{The first vertex of an edge.}
  \item{v2}{The second vertex of an edge.}
  \item{w}{The weight of the edge between v1 and v2.}
}
}
\usage{
wiki_graph
}
\description{
A data frame representing a weighted graph used for the Dijkstra
algorithm.
}
\references{
\url{https://en.wikipedia.org/wiki/Graph_(discrete_mathematics)}
}
\keyword{datasets}
