% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Calculate shortest paths using Dijkstra's algorithm}
\usage{
dijkstra(graph, init)
}
\arguments{
\item{graph}{A data frame containing the graph edges with variables
v1, v2, and w, where v1 is the starting vertex, v2 is the destination
vertex, and w is the edge weight.}

\item{init}{A numeric scalar specifying the initial vertex.

The algorithm calculates the shortest path from the initial vertex
to every other vertex in the graph.}
}
\value{
A numeric vector containing the shortest path from the initial
vertex to every other vertex.
}
\description{
Calculate shortest paths using Dijkstra's algorithm
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}

wiki_graph.R
}
