% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Calculate the greatest common divisor using the Euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric scalar or integer}

\item{b}{A numeric scalar or integer

The algorithm repeatedly calculates the remainder of a divided by b,
then replaces a with b and b with the remainder.
This process continues until b becomes zero.
The value of a is then the greatest common divisor.}
}
\value{
The greatest common divisor of a and b as a numeric value.
}
\description{
Calculate the greatest common divisor using the Euclidean algorithm
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
\url{https://www.khanacademy.org/computing/computer-science/cryptography/modarithmetic/a/the-euclidean-algorithm}
}
