% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki graph dataset}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{The starting vertex of the edge.}
  \item{v2}{The destination vertex of the edge.}
  \item{w}{The weight of the edge.}
}
}
\usage{
wiki_graph
}
\description{
A graph represented as an edge list, where each row represents
an edge from one vertex to another with an associated weight.
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}
}
\keyword{datasets}
